+++ standalone/inc/common/DMDWindow.cpp
@@ -4,7 +4,7 @@
 
 namespace VP {
 
-DMDWindow::DMDWindow(const std::string& szTitle, int x, int y, int w, int h, int z, int rotation)
+DMDWindow::DMDWindow(const string& szTitle, int x, int y, int w, int h, int z, int rotation)
     : VP::Window(szTitle, x, y, w, h, z, rotation)
 {
    m_destRect = { 0.0f, 0.0f, (float)w, (float)h };